requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
adapterremoval=2.3.3=pl5321h6dccd9a_2
anyio=3.7.0=pyhd8ed1ab_1
argon2-cffi=21.3.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py310h5764c6d_3
asttokens=2.2.1=pyhd8ed1ab_0
async-lru=2.0.2=pyhd8ed1ab_0
atk-1.0=2.38.0=hd4edc92_1
attrs=23.1.0=pyh71513ae_1
babel=2.12.1=pyhd8ed1ab_1
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
beautifulsoup4=4.12.2=pyha770c72_0
biopython=1.81=py310h1fa729e_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
bottleneck=1.3.7=py310h0a54255_0
bowtie2=2.5.1=py310h8d7afc0_0
brotli=1.0.9=h166bdaf_8
brotli-bin=1.0.9=h166bdaf_8
brotlipy=0.7.0=py310h5764c6d_1005
bzip2=1.0.8=h7f98852_4
c-ares=1.18.1=h7f8727e_0
ca-certificates=2023.11.17=hbcca054_0
cairo=1.16.0=hb05425b_4
certifi=2023.11.17=pyhd8ed1ab_0
cffi=1.15.1=py310h255011f_3
charset-normalizer=3.1.0=pyhd8ed1ab_0
cloudpickle=3.0.0=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.1.3=pyhd8ed1ab_0
contourpy=1.0.7=py310hdf3cbec_0
cryptography=39.0.1=py310h9ce1e76_0
cutadapt=4.4=py310h4b81fae_1
cycler=0.11.0=pyhd8ed1ab_0
cython=3.0.6=py310hc6cd4ac_0
debugpy=1.6.7=py310heca2aa9_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
dnaio=0.10.0=py310h4b81fae_3
docopt=0.6.2=py_1
entrypoints=0.4=pyhd8ed1ab_0
et_xmlfile=1.1.0=pyhd8ed1ab_0
exceptiongroup=1.1.1=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
expat=2.5.0=hcb278e6_1
fftw=3.3.10=nompi_hc118613_107
flit-core=3.9.0=pyhd8ed1ab_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.1=hef1e5e3_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.39.4=py310h2372a71_0
freetype=2.12.1=hca18f0e_1
fribidi=1.0.10=h36c2ea0_0
gdk-pixbuf=2.42.6=h04a7f16_0
gettext=0.21.1=h27087fc_0
giflib=5.2.1=h0b41bf4_3
glib=2.76.3=hfc55251_0
glib-tools=2.76.3=hfc55251_0
graphite2=1.3.14=h295c915_1
graphviz=3.0.0=h5abf519_1
gtk2=2.24.33=h90689f9_2
gts=0.7.6=h64030ff_2
harfbuzz=4.3.0=hf52aaf7_1
hisat2=2.2.1=hdbdd923_6
htslib=1.17=h6bc39ce_1
icu=58.2=hf484d3e_1000
idna=3.4=pyhd8ed1ab_0
importlib-metadata=6.6.0=pyha770c72_0
importlib_metadata=6.6.0=hd8ed1ab_0
importlib_resources=5.12.0=pyhd8ed1ab_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.23.1=pyh210e3f2_0
ipython=8.14.0=pyh41d4057_0
ipython_genutils=0.2.0=py_1
isa-l=2.30.0=ha770c72_4
jedi=0.18.2=pyhd8ed1ab_0
jinja2=3.1.2=pyhd8ed1ab_1
jpeg=9e=h0b41bf4_3
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.17.3=pyhd8ed1ab_0
jupyter-lsp=2.2.0=pyhd8ed1ab_0
jupyter_client=8.2.0=pyhd8ed1ab_0
jupyter_core=5.3.0=py310hff52083_0
jupyter_events=0.6.3=pyhd8ed1ab_0
jupyter_server=2.6.0=pyhd8ed1ab_0
jupyter_server_terminals=0.4.4=pyhd8ed1ab_1
jupyterlab=4.0.2=pyhd8ed1ab_0
jupyterlab_pygments=0.2.2=pyhd8ed1ab_0
jupyterlab_server=2.22.1=pyhd8ed1ab_0
k8=0.2.5=hdcf5f25_4
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.4=py310hbf28c38_1
krb5=1.20.1=hf9c8cef_0
lcms2=2.14=h6ed2654_0
ld_impl_linux-64=2.40=h41732ed_0
lerc=4.0.0=h27087fc_0
libblas=3.9.0=12_linux64_mkl
libbrotlicommon=1.0.9=h166bdaf_8
libbrotlidec=1.0.9=h166bdaf_8
libbrotlienc=1.0.9=h166bdaf_8
libcblas=3.9.0=12_linux64_mkl
libcurl=7.87.0=h6312ad2_0
libdeflate=1.13=h166bdaf_0
libedit=3.1.20221030=h5eee18b_0
libev=4.33=h516909a_1
libexpat=2.5.0=hcb278e6_1
libffi=3.4.2=h7f98852_5
libgcc-ng=13.1.0=he5830b7_0
libgd=2.3.3=h695aa2c_1
libgfortran-ng=13.1.0=h69a702a_0
libgfortran5=13.1.0=h15d22d2_0
libglib=2.76.3=hebfc3b9_0
libgomp=13.1.0=he5830b7_0
libhwloc=2.8.0=h32351e8_1
libiconv=1.17=h166bdaf_0
liblapack=3.9.0=12_linux64_mkl
libnghttp2=1.51.0=hdcd2b5c_0
libnsl=2.0.0=h7f98852_0
libpng=1.6.39=h753d276_0
librsvg=2.52.5=h0a9e6e8_2
libsodium=1.0.18=h36c2ea0_1
libsqlite=3.42.0=h2797004_0
libssh2=1.10.0=haa6b8db_3
libstdcxx-ng=13.1.0=hfd8a6a1_0
libtiff=4.4.0=h0e0dad5_3
libtool=2.4.7=h27087fc_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.2.4=h522a892_0
libwebp-base=1.2.4=h166bdaf_0
libxcb=1.15=h0b41bf4_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
libzlib=1.2.13=h166bdaf_4
llvm-openmp=16.0.5=h4dfa4b3_0
lxml=4.9.1=py310h5764c6d_0
lz4-c=1.9.4=hcb278e6_0
markdown-it-py=3.0.0=pyhd8ed1ab_0
markupsafe=2.1.3=py310h2372a71_0
matplotlib-base=3.7.1=py310he60537e_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mdurl=0.1.0=pyhd8ed1ab_0
minimap2=2.26=he4a0461_1
mistune=2.0.5=pyhd8ed1ab_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310ha2c4b55_0
mkl_fft=1.3.1=py310h2b4bcf5_1
mkl_random=1.2.2=py310h00e6091_0
munkres=1.1.4=py_0
nbclassic=1.0.0=pyhb4ecaf3_1
nbclient=0.8.0=pyhd8ed1ab_0
nbconvert=7.4.0=pyhd8ed1ab_0
nbconvert-core=7.4.0=pyhd8ed1ab_0
nbconvert-pandoc=7.4.0=pyhd8ed1ab_0
nbformat=5.9.0=pyhd8ed1ab_0
ncurses=6.4=hcb278e6_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
networkx=3.1=pyhd8ed1ab_0
notebook=6.5.4=pyha770c72_0
notebook-shim=0.2.3=pyhd8ed1ab_0
numexpr=2.8.4=py310h8879344_0
numpy=1.23.5=py310hd5efca6_0
numpy-base=1.23.5=py310h8e6c178_0
openpyxl=3.1.1=py310h1fa729e_0
openssl=1.1.1w=hd590300_0
overrides=7.3.1=pyhd8ed1ab_0
packaging=23.1=pyhd8ed1ab_0
pandas=2.0.2=py310h7cbd5c2_0
pandoc=2.19.2=h32600fe_2
pandocfilters=1.5.0=pyhd8ed1ab_0
pango=1.50.7=hbd2fdc8_0
parso=0.8.3=pyhd8ed1ab_0
patsy=0.5.3=pyhd8ed1ab_0
pbzip2=1.1.13=0
pcre=8.45=h9c3ff4c_0
pcre2=10.40=hc3806b6_0
perl=5.32.1=2_h7f98852_perl5
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pigz=2.6=h27826a3_0
pillow=9.4.0=py310h6a678d5_0
pip=23.1.2=pyhd8ed1ab_0
pixman=0.40.0=h36c2ea0_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_0
platformdirs=3.5.3=pyhd8ed1ab_0
pooch=1.7.0=pyha770c72_3
prometheus_client=0.17.0=pyhd8ed1ab_0
prompt-toolkit=3.0.38=pyha770c72_0
prompt_toolkit=3.0.38=hd8ed1ab_0
psutil=5.9.5=py310h1fa729e_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.15.1=pyhd8ed1ab_0
pygraphviz=1.9=py310h7dd584d_2
pynvml=11.4.1=pyhd8ed1ab_0
pyopenssl=23.2.0=pyhd8ed1ab_1
pyparsing=3.0.9=pyhd8ed1ab_0
pyrsistent=0.19.3=py310h1fa729e_0
pysam=0.21.0=py310hff46b53_0
pysocks=1.7.1=pyha2e5f31_6
python=3.10.10=h7a1cb2a_2
python-dateutil=2.8.2=pyhd8ed1ab_0
python-fastjsonschema=2.17.1=pyhd8ed1ab_0
python-isal=1.1.0=py310h2372a71_2
python-json-logger=2.0.7=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd8ed1ab_0
python_abi=3.10=2_cp310
pytz=2023.3=pyhd8ed1ab_0
pyyaml=6.0=py310h5764c6d_5
pyzmq=25.1.0=py310h5bbb5d0_0
readline=8.2=h8228510_1
requests=2.31.0=pyhd8ed1ab_0
rfc3339-validator=0.1.4=pyhd8ed1ab_0
rfc3986-validator=0.1.1=pyh9f0ad1d_0
rich=13.7.0=pyhd8ed1ab_0
samtools=1.17=hd87286a_1
scalene=1.5.31.1=py310hc6cd4ac_2
scipy=1.10.1=py310ha4c1d20_3
seaborn=0.13.0=hd8ed1ab_0
seaborn-base=0.13.0=pyhd8ed1ab_0
send2trash=1.8.2=pyh41d4057_0
seqkit=2.4.0=h9ee0642_0
setuptools=67.7.2=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.0=pyhd8ed1ab_0
soupsieve=2.3.2.post1=pyhd8ed1ab_0
sqlite=3.42.0=h2c6b66d_0
stack_data=0.6.2=pyhd8ed1ab_0
statsmodels=0.14.0=py310h278f3c1_1
tbb=2021.7.0=h924138e_1
terminado=0.17.1=pyh41d4057_0
testpath=0.6.0=pyhd8ed1ab_0
tinycss2=1.2.1=pyhd8ed1ab_0
tk=8.6.12=h27826a3_0
tomli=2.0.1=pyhd8ed1ab_0
tornado=6.3.2=py310h2372a71_0
tqdm=4.65.0=pyhd8ed1ab_1
traitlets=5.9.0=pyhd8ed1ab_0
typing-extensions=4.6.3=hd8ed1ab_0
typing_extensions=4.6.3=pyha770c72_0
typing_utils=0.1.0=pyhd8ed1ab_0
tzdata=2023c=h71feb2d_0
unicodedata2=15.0.0=py310h5764c6d_0
urllib3=1.26.14=py310h06a4308_0
wcwidth=0.2.6=pyhd8ed1ab_0
webencodings=0.5.1=py_1
websocket-client=1.5.3=pyhd8ed1ab_0
wheel=0.40.0=pyhd8ed1ab_0
xopen=1.7.0=py310hff52083_1
xorg-libxau=1.0.11=hd590300_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.10=h5eee18b_1
yaml=0.2.5=h7f98852_2
zeromq=4.3.4=h9c3ff4c_1
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=h166bdaf_4
zstandard=0.19.0=py310hdeb6495_1
zstd=1.5.2=h3eb15da_6
